# api/openapi.yaml
openapi: 3.0.3
info:
  title: Entregador API
  version: 1.0.0
  description: |
    API da portaria: moradores, entregas e notificações por WhatsApp.

    Todas as rotas `/v1` exigem um ID token do Firebase Auth (`Authorization: Bearer <token>`),
    exceto quando a API roda localmente com `AUTH_ENABLED=false`.

    Papéis: `admin` (síndico) pode tudo; `doorman` (porteiro) registra, retira e consulta entregas
    e consulta moradores e apartamentos.
servers:
  - url: http://localhost:8081
    description: Local
security:
  - firebaseAuth: []
...
paths:
...
  /internal/pubsub/notifications:
    post:
      summary: Push do Pub/Sub (uso interno)
      description: |
        Recebe o `NotifyDeliveryCommand` pela push subscription. Não é para o front.
        Em produção exige o token OIDC assinado pela service account do push.
      operationId: receivePubSubPush
      tags: [Interno]
      security:
        - pubSubPushToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PubSubPush"
      responses:
        "204":
          description: Mensagem processada (ack), ou tipo de evento ignorado.
        "400":
          description: Mensagem malformada; vai para o dead-letter depois das tentativas.
        "401":
          description: Token do push ausente ou inválido.
        "500":
          description: Falha ao notificar; o Pub/Sub tenta de novo.
components:
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ID token do Firebase Auth. O papel vem na claim `role`.
    pubSubPushToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token OIDC que o Pub/Sub assina com a service account do push.
...
  schemas:
...
    PubSubPush:
      type: object
      required: [message]
      properties:
        message:
          type: object
          properties:
            data:
              type: string
              format: byte
              description: NotifyDeliveryCommand em JSON, codificado em base64.
            attributes:
              type: object
              additionalProperties:
                type: string
            messageId:
              type: string
        subscription:
          type: string
        deliveryAttempt:
          type: integer